feat(SFINT-3353): Refactor + remove all notion of impression from SVC for now
